pyproject.toml
[project]
description = "Simple API deployed using AWS Lambda and API Gateway deployed using CDK"
name = "simple-api-lambda-cdk"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

# Core dependencies (always installed)
dependencies = []

[project.scripts]
dev = "app.handler:dev"
start = "app.handler:start"

# Install both "lambda" and "dev" groups by default when running `uv sync`
[tool.uv]
default-groups = "all"

[dependency-groups]
# Lambda runtime dependencies
lambda = [
    "requests>=2.31.0",
    "fastapi>=0.115.0",
    "mangum>=0.17.0",
    "uvicorn>=0.37.0",
    "scalar-fastapi>=1.4.3",
    "types-requests>=2.32.4.20250913",
    "pydantic[email]",
]

# Development dependencies
dev = [
    "pylint>=3.0.0",
    "mypy>=1.8.0",

    # CDK for local development
    "aws-cdk-lib>=2.220.0",
    "constructs>=10.4.2",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
# The directories to actually build (we only need the cdk directory)
packages = ["cdk"]

[tool.hatch.version]
path = "app/version.py"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

# Performance optimizations so mypy isn't so slow
cache_dir = ".mypy_cache"
incremental = true
sqlite_cache = true
skip_version_check = true

# Exclude generated and build directories
exclude = [
    "cdk/",
    "cdk.out/",
    ".venv/",
    "__pycache__/",
    ".*\\.pyc$",
    "build/",
    "dist/",
    ".git/",
]
